requirements.txt
# Confluence Wiki MCP 依赖包
atlassian-python-api>=3.41.0
mcp>=1.0.0
python-dotenv>=1.0.0
